Testing TCP Client-Server in Rust Found a crate to easily parse HTTP Requests Messing with the code to better understand it Training for the DogeGuard project
