+++ rust/tcp-client/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "tcp-client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+http_request_parser = "0.2.0"
